Only set top/bottom pattern to concentric for initial layer
I'm taking responsibility for this. I'm changing it as a reaction to #4167 since I think concentric is good for the first layer but not so good for the rest of the layers. That user seems to agree but it hasn't been tested thoroughly. It's my intuition, sorry if it went wrong.