Fix adhesion type checkbox if default is skirt
The remove-unused-values feature doesn't work with settings that are resolved, because we try to remove them at the appropriate position in the stack but then we check at that position in the stack if the value is not the same as what it was, but with this check the resolve function is no longer evaluated. It's complex, but hard to solve because we can't try to remove it at this location and predict that this will change the actual final value after the resolve function has done its thing.

Contributes to issue CURA-5877.