Fix translation mistakes found in Cura 3.6
These mistakes were found in the translations. In many cases it makes the text closer to the original rather than more consistent with the environment. If the environment is inconsistent then we should fix the original string in our source code rather than in the translations.

Contributes to issue CURA-5870.