Improve generic profiles for fabtotum and zyyx
These printers had profiles for zyyx_pro and fabtotum material suites, which were specific to those materials, but caused the generic materials to no longer have any profile. Let's use these machine-specific profiles for the generic materials as well. For the specific materials this changes nothing because the same profiles are still matched. For the generic materials this should improve the quality if the quality profile is tuned in any way to the machine.