Don't show custom/auto toggle buttons when not connected
In order to prevent suddenly switching to the other side when the connection is lost or restored, we only evaluate this upon opening the popup. This way you might be surprised that closing and then opening it can show something else, but it will never surprise you while working on the popup itself as a user.

Contributes to issue CURA-5876.