Make icon height equal to text height
This seems to be about the same actually as what it was. But now the relation is in there.
I also made the right icon inherit as much as possible from the left icon so that if we change something, we only have to change it in the left icon.

Contributes to issue CURA-5876.