Simplify sectionHeading element
No need to define that component in a higher location, just define it where you need it. Also, no need to use a Rectangle for this if it has no colour, just use an Item, but the Item itself is also not necessary if you just need the padding.

Contributes to issue CURA-5876.