Correct colours for the selected configuration
This should always be the same as the ExpandableComponent's background, so white when it's not active and light blue when it's active, regardless of theme.
The naming of this theme entry is a bit weird because the ActionButton theme is used for ExpandableComponent and now for ConfigurationItem as well, but that's why we should NEVER name these theme entries to something specific to one item. Oh well, I'm not about to refactor that now with all the branches going on everywhere.

Contributes to issue CURA-5876.