Change default profile of Aurora and Alfawise printers
Instead of the default layer height, we should change the default quality profile.
This is necessary because the 'normal' quality profile doesn't define a layer height, so that should inherit from the 0.1mm default layer height. But if the printer turned the default into a 0.15mm layer height then that is wrong.
Maybe we should let the normal quality profile overwrite it to 0.1mm, always?

Contributes to issue CURA-5902.