Don't show materials if printer has no materials
But if the printer does have other configurations to change, do show a placeholder text to indicate that the configuration can be selected here.
This also simplifies a bit of code where it would need to call an updateEnabled() function, since it turns out that these properties in Cura.MachineManager have proper signals (contrary to what was previously used, the metadata entry stuff).

Contributes to issue UCRA-5876.