Remove padding from views selector
This padding made it break the height of the item for some reason. I guess that is a Qt bug.
In any case, the padding shouldn't be there either since the lining is on the inside of the child buttons so the padding only causes a gap to appear on the left and right sides, which looks weird.

Contributes to issue CURA-6029.