Fix bold fonts on Linux
I'm not sure how this performs on other platforms, but on mine it needs at least 63 weight for bold. This is because there is a normal font at weight 40 and a bold font at weight 85. From 63+ it rounds to the bold font with the closest weight then.

Contributes to issue CURA-6025.