Fix type error
The typing was added by Nallath just now but it is giving an error for me.
The private access is allowed here because we're implementing it in the same class, even though it is a different instance.

This fixes the type error a bit more strictly than what lfei just did at the same time as me.