When disabling an extruder due to sync, always show a message.
Originally it would only show a message if the extruder was actually used,
but this didn't take the flow into account when you would first sync and then load a model.
In that case, the user never got a message and never got any notification that (and why) any of the extruders
got disabled.

CURA-5693