Add the discovered printer functionality
This can be used in the add machine wizard to short circuit the creation of a printer.
So instead of first selecting a printer and then connecting it, it will be possible to have
a list that contains all network discovered printers (from all plugins that can do this).

This means we can select one of those printers and directly add it (thus no longer needing that step)

CURA-6179