Increment setting_version
This must be up-to-date with the setting version that Cura currently uses, in order to ensure that the settings have the same names and the same functionality.
The settings that are set by these profiles are all nozzle-dimensional settings that haven't changed since 3.5, so we don't need to actually change the settings right now.

Contributes to issue CURA-6262.