Translate 'on' and 'off' in stage menu
It needs to be short. I used Google Translate for this. I translated the words 'active' and 'inactive' since this is about support being enabled or disabled. However with the translation of the word 'disabled' I'm getting results that seem to go more towards 'handicapped' rather than 'deactivated' so I think this is better.

Contributes to issue CURA-6257.