Fix welcome page buttons text not fitting
On some operating systems the text would not fit due to different font rendering. Also, after translation, it probably wouldn't fit either.
Usually we have the button size adjusting to the text inside it. The design didn't do this, either because it was too difficult to do in Zeplin, because the designer was lazy, or because the designer didn't know that this was the normal style in the rest of Cura. And the programmer took that over literally.

Contributes to issue CURA-6057.