Remove welcome_pages_default_margin theme entry
Use the wide margin instead. It's practically the same, nobody is going to notice. It still looks spacey and nice, and it's more consistent with the rest of the interface.
We had previously agreed that we will not make any new widget-specific theme entries any more.

Contributes to issue CURA-6057.