Reduce support horizontal expansion to 0
We have a long-standing issue in CuraEngine that support can be thinner than one line width and is then not generated, but if there is later some extra room it still gets supported in lower layers. This results in unnecessary support.

That issue in itself is a bit hairy to fix. But it does point out to another weirdness: PLA support has a horizontal offset of 0.2mm. This was introduced when the Support Horizontal Expansion setting was made as the default, but it was never really investigated. I asked Ultimaker's process engineers and we agreed that the setting could be set to 0 except for PVA.

Contributes to issue CURA-6438.