Make minimum wall flow also settable per extruder
It was settable per mesh but not per extruder, but we want it to be settable per extruder as well, causing all meshes printed with that extruder to use that setting value. That's how most of these settings work (like Outer Wall Speed and such).
There are some exceptions to this where a setting only makes sense if applied in per-object settings, such as Infill Mesh or support blockers, but this is not such a setting.

Fixes #5622.