Set maximum feedrate to light speed
For many printers this maximum feedrate is unknown. The rationale for this commit is that if it's unknown we shouldn't limit the speed settings.
This does cause the time estimations to be out of whack if the speed is limited on one axis but not on another axis, such as if you home the build plate while homing the head as well (when the head is much faster than the build plate). Because now Cura doesn't assume that there is a limit on any axis any more. It just holds the print speeds.
Many printers override these settings so that they have a more realistic value. Previously the value of 500mm/s wasn't very realistic anyway, as most printers have somewhere in the area of 250mm/s in their firmware limits.

Fixes #5640.