Cast autosave delay preference to int
Oops. How long has this been in there, you think? At least 3 years.
As seen in #5729, when changing this preference it results in an error that it can't set the interval to a string. The default is an integer, so if you don't edit it it works fine, but if you edited the preferences file, it gets a string from the preferences file. Casting both to an int should work.