Add new material properties for cleanly breaking off materials
When performing a long retract, the firmware can use these material properties to determine how fast and how far it needs to retract.
The retraction goes in two stages: Anti-ooze retraction stops oozing, and then the break retraction breaks off the filament from the bead in the nozzle.

Contributes to issue CURA-6329.