Break up translation in two pieces
As a rule of thumb, translated text should not contain any layout. This particular text has gone wrong in the translations because the translators didn't translate the second part (or their application messed it up). I'm breaking it up into two pieces as to force them to translate both.

Contributes to issue CURA-6547.