Fix translatability of caution message
Never put the formatting within the i18nc call. Always put the formatted string inside, but the formatting call outside of the function, like this:

    catalog.i18nc('@info', 'I am {age} years old.').format(age = my_age)

Otherwise, when looking up the translation for the string, it's going to try to look up the translation for the formatted string (with the age already filled in). It won't be able to find that because this sentence was not translated for all possible ages of a human.
In this case I can make it even easier on the translator because the list must always follow after the text.