Fix the updating of material diameter.
This was a weeeird bug. It worked if you changed the value and selected another field.
Pressing enter however (aka; The behavior that 80%+ of the users have) would change it
but it would switch back in half a second. It seems to have been caused by mutliple onEditingFinished
signals being emitted in sequence.
As I can't be bothered to go through the bowels of QML and figure out why exactly this is doing what it's doing,
I just tied the enter to losing the focus, which ensures that the right behavior happens (Field is no longer selected,
and the onEditingFinished is emitted)

Contributes to CURA-6590