Really don't make creality_base a visible printer
It's not a printer. It's a base profile. Every printer in the list of printers should coincide with a physical printer, with the one exception of Custom FFF Printer.
If we find settings that work for all printers we should apply it to fdmprinter so that we do apply it to every printer.

Contributes to issue CURA-6554.