Don't set a formula that'll evaluate to 'on' for randomize infill start.
It's policy to default to the old behaviour as much as possible for new features. We can maybe default to true (for these values) if we ask the materials team later. Instead opted to warn the user if the infill isn't in the list of those patterns.

part of CURA-6463