Nix 3mf-read signal-postphoning, which prevented connects.
This fix was originally made to fix CURA-5056, but it turns out it now gives warnings and possibly messes up 3mf reading instead of fixing it, since it prevents the connection of new listeners to those signals.
After removal, the original file this was reported against still loads instead of the crash it was reported to give in the ticket (reproduction rate supposedly 100%) so the removal probably doesn't mess things up.

part of CURA-6600