Add separate definition for Ultimaker 2 with Olsson block
This is necessary because our ContainerTree class can only create one tree per definition. This tree is not created separately for every added printer and as such the tree can't be different for every added printer.
However the Ultimaker 2 Olsson block upgrade allows selecting extra variants for the UM2. This changes the available variants in the container tree, so this would be impossible.

By making it a separate definition, it now gets a separate tree. However this does mean that we need to do a version upgrade for this as well.

Contributes to issue CURA-6775.