plugins: Reweighting stages
In theory it should be possible to add stages, but there is no possiblity to add any other stage in between. Therefore I moved the weights by 10+n*10. So we can add 10 before prepare and 9 stages between those 3 known stages.
It is probably possible to set the same weight to multiple stages, but I'm not sure how Cura decides, which stage to take first then. By the moment the plugin's init is called? Or by alphabet?
Therefore putting some space in between the weights should give some clearity.