Make fixes found through manual checking of translations
I'm not a native speaker of French or Portuguese but enough to make those corrections of not translating 'Trimesh'. I'm not an expert of the Korean language to fix that missing translation but I made do with someone who learnt it as a second language. I am an expert at Dutch so I made a couple of fixes there.

Contributes to issue CURA-6768.