Add StartupWMClass=cura.real in order to group windows properly in Gnome and more.
This has been infuriating me for the past few weeks. The result of this change is that when you use Cura as built from source or via thopiekar's PPA the Gnome shortcut will act only as a launcher for potentially infinite instances of Cura. That is, if you keep clicking it, it will open up more and more instances of Cura. This is because the class cura.real is not present or specified in the desktop file. See this askubuntu thread on the topic;

https://askubuntu.com/questions/367396/what-does-the-startupwmclass-field-of-a-desktop-file-represent/367851#367851

Simply adding "StartupWMClass=cura.real" allows Cura to behave like a normal application in all desktop environments that implement .desktop. Cura can now spawn children and in gnome in particular has options for spawning a New Window, rather than a runaway "Spawn more instances" as you click the launcher. This has been particularly annoying for me, as in a makerspace I attend where I have set up a dedicated Cura machine people have been opening several instances of Cura because of this flaw.