Creality Legacy Printer Definitions
The major changes to the Creality printer definitions has caused quite a stir in the miniature hobbyist forums and Facebook groups. Importing old profiles does not give the end-user the expected results due to the previous template profiles no longer available to the new Creality printer definitions. This is my first step into helping that community get their previous qualities and definitions back.

Note, apparently there are enough differences between the legacy Creality definitions and the plain custom fdmprinter definitions to make a significant difference.

I've given these files to a couple of Facebook groups and they are receiving a lot of accolades.

My thought is if the powers that be are okay with approving this, we can work to integrate them into the same manufacturer for the next release, instead of creating a second "Creality Legacy 3D" manufacturer.

PS Daniel Herrero says he discussed this with @nallath , which is why I'm submitting this for a pull. This is the first time I've ever used GitHub to contribute as I'm not really a programmer.