Fix overriding of prime tower position
We had changed it for all printers back when we changed the prime tower position to a formula, but since then we've been getting new printer definitions that were made from templates floating around with the old structure.
We should really have an automated test for this.

Discovered during review of #6518.