Sort definition files before testing them
When going through the list of tests for many printers to fix them, it's then easier to go through the failed list. They will appear in the same order as in my IDE. Just a bit of an ease of life thing without any real cost.

Done during Turbo Testing & Tooling.