Remove unused default_value overrides
These are overrides of default_value while there is a value defined. As such these overrides had no effect at all.
Changing them to value can actually change the behaviour of the profile. That is not what the profile author has apparently tested with, so I'm not doing that. I'm just removing the unused data.
In the future if we get new definitions the author gets notified of the error so that he may test more effectively. The legacy can't be fixed without re-testing, so I'm leaving that to the authors.

Done during Turbo Testing & Tooling. I want to go home for the weekend now...