Only show conical support min width if angle is positive
If the angle is 0 or negative, this setting won't have any effect since the support cannot be made wider than it already is using this setting. It can only prevent it from becoming smaller, and it will only become smaller from a positive conical support angle.

Discovered during work on the Settings Guide.