Remove all overrides of Z Hop Enabled
So the Z hop has been broken for a while, such that it would not hop any more except for extruder switches. We've fixed that now, but that means that a lot of printers would suddenly hop where they didn't hop before. This could degrade print quality and productivity for a lot of printers. So we're now disabling Z hops for every printer. The result should be the same. We're making Z hops for extruder switches separate from the normal Z hop setting so that can still be enabled.
For profiles where Z hop was overridden to be False, that was useless to begin with so I've removed those as well.

Contributes to issue CURA-6860.