Make small feature max length a child setting of small hole max size
The small hole max size is not used in CuraEngine. It may not be a leaf setting then. Instead, Small Feature Max Length is used which is directly dependent on Small Hole Max Size. They are simply different units of the same thing. This is how we should structure these settings then.