Fix skip (or late) pause at height.
Previously, the line after the layer change would also contain the z height.
If not found, the could would break the per-line reading, and skip to the next layer.
However, if the code doesn't contain the Z on the first gcode line of the layer, this would cause it to skip past all layers, until there was a layer with the Z in the first line (so this also explains the 'late' pauses).
I just ignored the optimization and made it a continue.
I checked, and I don't think there's any way this causes the postproc.-gcode to be inserted twice, since it returns the complete data after insering it once.

CURA-6965