Increase printer size and add disallowed areas
With the stock firmware of the Ender 3, the size of the original printer was wrong, causing the model to land in a different place on the build plate. This led people to believe that the printer could take bigger prints. However it couldn't; it was just not allowing to print all the way down to the coordinate origin at 0,0.

This makes the printer bigger but also adds disallowed areas so that you can't print anything near 0,0.

Fixes #6267.