Make template a bit less intimidating and restore brackets again
I think the 'why' of things is not always relevant and makes people prevent reading all of it. Keep it brief! And with all of this information that people tend to just leave there it's hard to find the actual changes that people made to the template. That's also why the brackets are important, so that it's easy to see that people didn't erase the original template text.
We used to have HTML comment codes there but then people put the actual bug report inside the comments as well and nothing showed up, so this is somewhat of a middle ground that everyone understands, programmers and non-programmers.