Remove useless override of bed temperature
The Generic PETG material also sets it to 70, as well as the eMotionTech and TiXYZ PETGs. The iMade3D bed temperature was 65, so for that one the temperature will change to 65 then which seems appropriate. I'll change the iMade3D thing in fdm_materials so that the settings are still the same.

Discovered during investigation of #6898.