Fix button to open in-Cura Marketplace materials
The button was opening the online Marketplace, which was wrong.
This fix changes the behavior to open the in-Cura Marketplace at
the materials tab.

In order to achieve that, the button calls an action called
materialsMarketplace, which in turn is connected through a
Connection in the ApplicationMenu. When the action is triggered,
it first launched the Toolbox and then it sets its category to
the materials tab.

Since the callExtensionMethod does not allow us to provide input
arguments to the called method, a new method had to be created in
the toolbox that changes the view to the materials tab, instead
of immediately calling the setViewCategory("material").

CURA-7027